polish the feature
* strip 3rd party dependency
* refactoring
