+++ mapper/lib/src/errors.dart
@@ -1,7 +1,4 @@
 import 'package:dart_json_mapper/dart_json_mapper.dart';
-import 'package:string_similarity/string_similarity.dart';
-
-import './model/index.dart';
 
 abstract class JsonMapperError extends Error {}
 
@@ -11,20 +8,18 @@ abstract class JsonFormatError extends JsonMapperError {
 }
 
 class JsonMapperSubtypeError extends JsonMapperError {
-  JsonMapperSubtypeError(this.discriminatorValue, this.validDiscriminators, this.superclass);
+  JsonMapperSubtypeError(
+      this.discriminatorValue, this.validDiscriminators, this.superclass);
 
   final dynamic discriminatorValue;
   final List<dynamic> validDiscriminators;
   final ClassInfo superclass;
 
   @override
-  String toString() {
-    validDiscriminators.sort((a, b ) => a.toString().similarityTo(discriminatorValue.toString()).compareTo(b.toString().similarityTo(discriminatorValue.toString())));
-    return '$discriminatorValue is not a valid discriminator value for ${superclass.reflectedType}. Did you mean ${validDiscriminators.last}?\n\n'
-        'Valid values are:\n${validDiscriminators.join(',\n')}';
-  }
+  String toString() =>
+      '"$discriminatorValue" is not a valid discriminator value for type "${superclass.reflectedType}".\n\n'
+      'Valid values are:\n${validDiscriminators.map((x) => '  - $x').join(',\n')}';
 }
-
 
 class _JsonFormatErrorImpl extends JsonMapperError implements JsonFormatError {
   final DeserializationContext _context;
