+++ mapper/lib/src/mapper.dart
@@ -9,7 +9,6 @@ import 'package:reflectable/reflectable.dart'
         DeclarationMirror,
         VariableMirror,
         MethodMirror;
-import 'package:string_similarity/string_similarity.dart';
 
 import 'errors.dart';
 import 'model/index.dart';
@@ -351,11 +350,11 @@ class JsonMapper {
           return _getTypeInfo(value);
         }
 
-        if(_discriminatorToType[value] == null) {
-          var subs = getAllSubTypes(objectClassInfo);
-          // Map each [ClassInfo] to its discriminator value
-          var validDiscriminators = subs.map((e) => e.getMeta(context.options.scheme)!.discriminatorValue).toList();
-
+        if (_discriminatorToType[value] == null) {
+          final validDiscriminators = ClassInfo.getAllSubTypes(
+                  _classes, objectClassInfo)
+              .map((e) => e.getMeta(context.options.scheme)!.discriminatorValue)
+              .toList();
           throw JsonMapperSubtypeError(
             discriminatorValue,
             validDiscriminators,
@@ -373,21 +372,6 @@ class JsonMapper {
           : typeInfo;
     }
     return typeInfo;
-  }
-
-  /// Returns all subtypes of [classInfo], this is transitive
-  List<ClassInfo> getAllSubTypes(ClassInfo classInfo) {
-    final result = <ClassInfo>[];
-    for (var subType in _classes.values) {
-      try {
-      if (
-          subType.classMirror.isSubtypeOf(classInfo.classMirror)) {
-        result.add(subType);
-      } } catch (e) {
-        // Ignore no capability to check subtype
-      }
-    }
-    return result;
   }
 
   Type? _getTypeByStringName(String? typeName) =>
