+++ mapper/lib/src/utils.dart
@@ -217,6 +217,22 @@ class ClassInfo {
     return null;
   }
 
+  /// Returns all subtypes of [classInfo], this is transitive
+  static List<ClassInfo> getAllSubTypes(
+      Map<Type, ClassInfo> classes, ClassInfo classInfo) {
+    final result = <ClassInfo>[];
+    for (final subType in classes.values) {
+      try {
+        if (subType.classMirror.isSubtypeOf(classInfo.classMirror)) {
+          result.add(subType);
+        }
+      } catch (e) {
+        // Ignore no capability to check subtype
+      }
+    }
+    return result;
+  }
+
   MethodMirror? getJsonAnySetter([dynamic scheme]) =>
       getJsonSetter(null, scheme);
