+++ models/unit_testing/lib/unit_testing.dart
@@ -2,6 +2,7 @@ export './generic.dart';
 export './immutable.dart';
 export './json.dart';
 export './model.dart';
+export './inheritance.dart';
 export 'unit_testing.mapper.g.dart'
     show
         unitTestingGeneratedAdapter,
