+++ perf-test/test/unit/test.errors.dart
@@ -1,6 +1,6 @@
 import 'package:dart_json_mapper/dart_json_mapper.dart';
 import 'package:test/test.dart';
-import 'package:unit_testing/unit_testing.dart' show Car, Color;
+import 'package:unit_testing/unit_testing.dart' show Car, Color, Business;
 
 class UnAnnotated {}
 
@@ -158,6 +158,24 @@ void testErrorHandling() {
     test('Missing target type for deserialization', () {
       expect(catchError(() => JsonMapper.deserialize('{}')),
           TypeMatcher<MissingTypeForDeserializationError>());
+    });
+
+    test('Discriminator property value is wrong', () {
+      // given
+      final json = '{"type":"p","a":1}';
+
+      // when
+      final error = catchError(() => JsonMapper.deserialize<Business>(json));
+
+      // then
+      expect(
+          error.toString(),
+          '"p" is not a valid discriminator value for type "Business".\n'
+          '\n'
+          'Valid values are:\n'
+          '  - BusinessType.private,\n'
+          '  - BusinessType.public');
+      expect(error, TypeMatcher<JsonMapperSubtypeError>());
     });
   });
 }
